+++ src/main/kotlin/com/cjfc/recipesmanager/presentation/payload/RecipeBasePayload.kt
@@ -1,0 +1,5 @@
+package com.cjfc.recipesmanager.presentation.payload
+
+data class RecipeBasePayload(
+    val id: String
+)